+++ lib/nadia/model.ex
@@ -7,8 +7,43 @@ defmodule Nadia.Model do
   """
 
   defmodule User do
-    defstruct id: nil, first_name: nil, last_name: nil, username: nil
-    @type t :: %User{id: integer, first_name: binary, last_name: binary, username: binary}
+    defstruct id: nil,
+              is_bot: nil,
+              first_name: nil,
+              last_name: nil,
+              username: nil,
+              language_code: nil,
+              is_premium: nil,
+              added_to_attachment_menu: nil,
+              can_join_groups: nil,
+              can_read_all_group_messages: nil,
+              supports_guest_queries: nil,
+              supports_inline_queries: nil,
+              can_connect_to_business: nil,
+              has_main_web_app: nil,
+              has_topics_enabled: nil,
+              allows_users_to_create_topics: nil,
+              can_manage_bots: nil
+
+    @type t :: %User{
+            id: integer,
+            is_bot: boolean,
+            first_name: binary,
+            last_name: binary,
+            username: binary,
+            language_code: binary,
+            is_premium: boolean,
+            added_to_attachment_menu: boolean,
+            can_join_groups: boolean,
+            can_read_all_group_messages: boolean,
+            supports_guest_queries: boolean,
+            supports_inline_queries: boolean,
+            can_connect_to_business: boolean,
+            has_main_web_app: boolean,
+            has_topics_enabled: boolean,
+            allows_users_to_create_topics: boolean,
+            can_manage_bots: boolean
+          }
   end
 
   defmodule ChatPhoto do
@@ -37,8 +72,15 @@ defmodule Nadia.Model do
   end
 
   defmodule PhotoSize do
-    defstruct file_id: nil, width: nil, height: nil, file_size: nil
-    @type t :: %PhotoSize{file_id: binary, width: integer, height: integer, file_size: integer}
+    defstruct file_id: nil, file_unique_id: nil, width: nil, height: nil, file_size: nil
+
+    @type t :: %PhotoSize{
+            file_id: binary,
+            file_unique_id: binary,
+            width: integer,
+            height: integer,
+            file_size: integer
+          }
   end
 
   defmodule Audio do
@@ -169,28 +211,70 @@ defmodule Nadia.Model do
 
   defmodule Message do
     defstruct message_id: nil,
+              message_thread_id: nil,
               from: nil,
+              sender_chat: nil,
+              sender_boost_count: nil,
+              sender_business_bot: nil,
+              sender_tag: nil,
               date: nil,
+              guest_query_id: nil,
+              business_connection_id: nil,
               chat: nil,
               forward_from: nil,
               forward_from_chat: nil,
+              forward_origin: nil,
               forward_date: nil,
               reply_to_message: nil,
+              external_reply: nil,
+              quote: nil,
+              reply_to_story: nil,
+              reply_to_checklist_task_id: nil,
+              reply_to_poll_option_id: nil,
+              via_bot: nil,
+              guest_bot_caller_user: nil,
+              guest_bot_caller_chat: nil,
               edit_date: nil,
+              has_protected_content: nil,
+              is_topic_message: nil,
+              is_automatic_forward: nil,
+              is_from_offline: nil,
+              is_paid_post: nil,
+              media_group_id: nil,
+              author_signature: nil,
+              paid_star_count: nil,
               text: nil,
               entities: nil,
+              link_preview_options: nil,
+              suggested_post_info: nil,
+              effect_id: nil,
+              animation: nil,
               audio: nil,
               document: nil,
+              live_photo: nil,
+              paid_media: nil,
               photo: [],
+              story: nil,
               sticker: nil,
               video: nil,
+              video_note: nil,
               voice: nil,
               caption: nil,
+              caption_entities: nil,
+              show_caption_above_media: nil,
+              has_media_spoiler: nil,
+              checklist: nil,
               contact: nil,
+              dice: nil,
+              game: nil,
+              poll: nil,
               location: nil,
               venue: nil,
               new_chat_member: nil,
+              new_chat_members: [],
               left_chat_member: nil,
+              chat_owner_left: nil,
+              chat_owner_changed: nil,
               new_chat_title: nil,
               new_chat_photo: [],
               delete_chat_photo: nil,
@@ -199,32 +283,76 @@ defmodule Nadia.Model do
               channel_chat_created: nil,
               migrate_to_chat_id: nil,
               migrate_from_chat_id: nil,
-              pinned_message: nil
+              pinned_message: nil,
+              reply_markup: nil,
+              web_app_data: nil
 
     @type t :: %Message{
             message_id: integer,
+            message_thread_id: integer,
             from: User.t(),
+            sender_chat: Chat.t(),
+            sender_boost_count: integer,
+            sender_business_bot: User.t(),
+            sender_tag: binary,
             date: integer,
+            guest_query_id: binary,
+            business_connection_id: binary,
             chat: Chat.t(),
             forward_from: User.t(),
             forward_from_chat: Chat.t(),
+            forward_origin: any,
             forward_date: integer,
             reply_to_message: Message.t(),
+            external_reply: any,
+            quote: any,
+            reply_to_story: any,
+            reply_to_checklist_task_id: integer,
+            reply_to_poll_option_id: binary,
+            via_bot: User.t(),
+            guest_bot_caller_user: User.t(),
+            guest_bot_caller_chat: Chat.t(),
             edit_date: integer,
+            has_protected_content: boolean,
+            is_topic_message: boolean,
+            is_automatic_forward: boolean,
+            is_from_offline: boolean,
+            is_paid_post: boolean,
+            media_group_id: binary,
+            author_signature: binary,
+            paid_star_count: integer,
             text: binary,
-            entities: MessageEntity.t(),
+            entities: [MessageEntity.t()],
+            link_preview_options: any,
+            suggested_post_info: any,
+            effect_id: binary,
+            animation: any,
             audio: Audio.t(),
             document: Document.t(),
+            live_photo: any,
+            paid_media: any,
             photo: [PhotoSize.t()],
+            story: any,
             sticker: any,
             video: any,
+            video_note: any,
             voice: any,
             caption: binary,
+            caption_entities: [MessageEntity.t()],
+            show_caption_above_media: boolean,
+            has_media_spoiler: boolean,
+            checklist: any,
             contact: any,
+            dice: any,
+            game: any,
+            poll: any,
             location: any,
             venue: any,
             new_chat_member: User.t(),
+            new_chat_members: [User.t()],
             left_chat_member: User.t(),
+            chat_owner_left: any,
+            chat_owner_changed: any,
             new_chat_title: binary,
             new_chat_photo: [PhotoSize.t()],
             delete_chat_photo: atom,
@@ -233,19 +361,33 @@ defmodule Nadia.Model do
             channel_chat_created: atom,
             migrate_to_chat_id: integer,
             migrate_from_chat_id: integer,
-            pinned_message: Message.t()
+            pinned_message: Message.t(),
+            reply_markup: any,
+            web_app_data: any
           }
   end
 
   defmodule MessageEntity do
-    defstruct type: nil, offset: nil, length: nil, url: nil, user: nil
+    defstruct type: nil,
+              offset: nil,
+              length: nil,
+              url: nil,
+              user: nil,
+              language: nil,
+              custom_emoji_id: nil,
+              unix_time: nil,
+              date_time_format: nil
 
     @type t :: %MessageEntity{
             type: binary,
             offset: integer,
             length: integer,
             url: binary,
-            user: User.t()
+            user: User.t(),
+            language: binary,
+            custom_emoji_id: binary,
+            unix_time: integer,
+            date_time_format: binary
           }
   end
 
@@ -278,18 +420,56 @@ defmodule Nadia.Model do
               message: nil,
               edited_message: nil,
               channel_post: nil,
+              edited_channel_post: nil,
+              business_connection: nil,
+              business_message: nil,
+              edited_business_message: nil,
+              deleted_business_messages: nil,
+              guest_message: nil,
+              message_reaction: nil,
+              message_reaction_count: nil,
               inline_query: nil,
               chosen_inline_result: nil,
-              callback_query: nil
+              callback_query: nil,
+              shipping_query: nil,
+              pre_checkout_query: nil,
+              purchased_paid_media: nil,
+              poll: nil,
+              poll_answer: nil,
+              my_chat_member: nil,
+              chat_member: nil,
+              chat_join_request: nil,
+              chat_boost: nil,
+              removed_chat_boost: nil,
+              managed_bot: nil
 
     @type t :: %Update{
             update_id: integer,
             message: Message.t(),
             edited_message: Message.t(),
             channel_post: Message.t(),
+            edited_channel_post: Message.t(),
+            business_connection: any,
+            business_message: Message.t(),
+            edited_business_message: Message.t(),
+            deleted_business_messages: any,
+            guest_message: Message.t(),
+            message_reaction: any,
+            message_reaction_count: any,
             inline_query: InlineQuery.t(),
             chosen_inline_result: ChosenInlineResult.t(),
-            callback_query: CallbackQuery.t()
+            callback_query: CallbackQuery.t(),
+            shipping_query: any,
+            pre_checkout_query: any,
+            purchased_paid_media: any,
+            poll: any,
+            poll_answer: any,
+            my_chat_member: any,
+            chat_member: any,
+            chat_join_request: any,
+            chat_boost: any,
+            removed_chat_boost: any,
+            managed_bot: any
           }
   end